First (possibly final) iteration.
This pretty much does the job. A first major improvement would
be to output three columns, not one (such that the column to the
right is a continuation of the column to the left, as in a newspaper.)

Also, column height would be enforced to be something like 48 (which
seems to work in Libreoffice.) So you'd end up with chunks that are
48 lines high, and that are situated next to each other, but the fourth
one belongs to a new set.

You'd likely need some data structure to fit the output of the
iterator, to enforce this topology of the output.